+++ includes/class-testimonials-widget-settings.php
@@ -897,15 +897,17 @@ class Testimonials_Widget_Settings extends Aihrus_Settings {
 
 		$defaults = self::get_defaults();
 
-		if ( ! empty( $input['has_archive'] ) )
+		if ( ! empty( $input['has_archive'] ) ) {
 			$input['has_archive'] = self::prevent_slug_conflict( $input['has_archive'] );
-		else
+		} else {
 			$input['has_archive'] = $defaults['has_archive'];
+		}
 
-		if ( ! empty( $input['rewrite_slug'] ) )
+		if ( ! empty( $input['rewrite_slug'] ) ) {
 			$input['rewrite_slug'] = self::prevent_slug_conflict( $input['rewrite_slug'] );
-		else
+		} else {
 			$input['rewrite_slug'] = $defaults['rewrite_slug'];
+		}
 
 		$flush_rewrite_rules = false;
 		// same has_archive and rewrite_slug causes problems
@@ -916,22 +918,25 @@ class Testimonials_Widget_Settings extends Aihrus_Settings {
 			$flush_rewrite_rules = true;
 		}
 
-		// did URL slugs change?
-		$has_archive  = tw_get_option( 'has_archive' );
-		$rewrite_slug = tw_get_option( 'rewrite_slug' );
-		if ( $has_archive != $input['has_archive'] || $rewrite_slug != $input['rewrite_slug'] )
+		// did URL slugs or taxonomy change?
+		$has_archive      = tw_get_option( 'has_archive' );
+		$rewrite_slug     = tw_get_option( 'rewrite_slug' );
+		$use_cpt_taxonomy = tw_get_option( 'use_cpt_taxonomy' );
+		if ( $has_archive != $input['has_archive'] || $rewrite_slug != $input['rewrite_slug'] || $use_cpt_taxonomy != $input['use_cpt_taxonomy'] ) {
 			$flush_rewrite_rules = true;
+		}
 
-		if ( $flush_rewrite_rules )
+		if ( $flush_rewrite_rules ) {
 			flush_rewrite_rules();
+		}
 
 		$input['version']        = self::$version;
 		$input['donate_version'] = Testimonials_Widget::VERSION;
 
 		$input = apply_filters( 'testimonials_widget_validate_settings', $input, $errors );
-		if ( empty( $do_errors ) )
+		if ( empty( $do_errors ) ) {
 			$validated = $input;
-		else {
+		} else {
 			$validated = array(
 				'input' => $input,
 				'errors' => $errors,